code review

+++ metadata-mis/src/main/java/com/mingdong/mis/domain/mapper/ApiReqMapper.java
@@ -9,6 +9,6 @@ public interface ApiReqMapper
 {
     void add(ApiReq o);
 
-    Long countBy(@Param("clientId") Long clientId, @Param("userId") Long userId, @Param("productId") Long productId,
+    int countBy(@Param("clientId") Long clientId, @Param("userId") Long userId, @Param("productId") Long productId,
             @Param("startTime") Date startTime, @Param("endTime") Date endTime);
 }
